Move the OverviewComponent into the Show namespace
