+++ app/components/show/collection_component.html.erb
@@ -9,7 +9,7 @@
 
 <div class="row mt-5">
   <div class="col-md-6">
-    <%= render OverviewComponent.new(presenter: presenter) %>
+    <%= render Show::OverviewComponent.new(presenter: presenter) %>
   </div>
   <div class="col-md-6">
     <%= render DetailsComponent.new(presenter: presenter) %>
