+++ spec/components/show/overview_component_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe OverviewComponent, type: :component do
+RSpec.describe Show::OverviewComponent, type: :component do
   let(:component) { described_class.new(presenter: presenter) }
   let(:presenter) { instance_double(ArgoShowPresenter, document: doc, cocina: cocina, change_set: change_set, state_service: state_service) }
   let(:change_set) { ItemChangeSet.new(cocina) }
